ml-service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
numpy
pydantic
python-dotenv
google-generativeai
pillow
